Acutally fix the crash when switching resolutions.
Crash is an access violation in native. Must be multi-threaded calling onDrawFrame separately from processFrame. "Fixed" by adding a synchronization. Trying to not hold the lock for longer than necessary, so only doing it on the resize. Seems to work, but with this kind of problem just might not be getting lucky.

Add a debug-ish view for eocvsim. This is a 2x2 grid of mats, currently showing the input and the mask for each of the colors.

Also using public statics, which eocvsim supports with UI elements so that those values can be updated live in the app. Note: Cannot also use @Config because it doesn't know how to import or compile that.

Grouped the debug visualization parts of the pipeline at the bottom, so that we don't need to do that part (and all the time it takes) when we are done.