Fix for getting the detected sample from the vision processor.
The main issue is pretty-much merge related. We started working in parallel, and Calvin had refactored the detection into a list of DetectedSample objects, instead of a list of centroid Point objects. From the TestVisionTeleop, we were getting the list of centroids, which was empty.

However, the fix is not simply to look at the detectedSamples list on the vision processor, but instead to add a method to get the detected sample. This enables two things. First, it means that the logic for detecting which sample to go after is the responsibility of the vision processor. No more .get(0) in the opmode. Second, it means that we can make the operation thread-safe. The processFrame() method is called from one thread, and the loop() method is called from another thread. We need to synchronize or coordinate whenever we access a variable which may be read or written from either thread.

For now, I added the "logic" of choosing which sample to return to be the largest sample (by area) that we found in the current frame.

Also:
* Adding toString() method on DetectedSample
* Adding FTCDashboard support. This conflicts with EOCV-Sim, so that's a pain. Examples show bitmap being RGB_565, but currently we are doing ARGB_8888. I am unable to test now so I left as a comment to remember to check it.
* Increase the saturation lower limit for red, to avoid skin tones
* The two threads is also a problem for telemetry. Doing telemetry from processFrame() is useful for EOCV-Sim when we are writing our pipeline in the first place, but just causes issues when we move to the Control Hub. We might be able to meet in the middle by using vision test opmodes in EOCV-Sim... Started to add a "doTelemetry()" method, which for now will just do the OpMode telemetry.
* Added a minimum area threshold to findCentroidsAndContours